requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
pillow==11.3.0
PyJWT==2.10.1
python-slugify==8.0.4
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
text-unidecode==1.3
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1